DEC: refactor into package structure
